fix: deep audit round 2 wave 2 — 61 remaining MEDIUM+LOW bugs fixed
Services (11):
- API keys moved from URL params to headers (giftasset WS, TonCenter)
- sendAgentTransaction wraps transfer in external message cell
- initTonDNS made async, appendDailyLog checks error code before fallback
- crew-system: namespace limit (50) + value size limit (10KB)
- image-service: os.tmpdir() instead of hardcoded /tmp
- telegram-gifts: removed dead _catalogFetch variable
- TonConnect: removed dead getUserSession, sessions TTL cleanup
- plugin-marketplace: SELECT FOR UPDATE on paid install
- telegram-userbot: 500ms rate limiter on all TG API calls

DB + orchestrator + execution-tools (11):
- agents.search() uses SQL ILIKE instead of in-memory filter
- agents.getStats() uses SQL COUNT aggregation
- userId=0 no longer bypasses ownership (changed to != null)
- DOUBLE PRECISION → NUMERIC(20,9) for financial amounts
- GetGems scraping marked deprecated with AbortSignal timeout
- Fragment auth now checks user ownership
- parseScheduleMs: added general каждые N минут/час patterns
- runningAgents periodic cleanup every 10min
- searchWeb deduped as alias of webSearch
- updateAgent name uniqueness via constraint catch

ai-agent-runtime (11):
- DNS rebinding documented as known limitation
- sanitizeForLog: 50K char truncation before regex
- mdToHtml: italic regex excludes HTML tags
- crypto moved to module-level import
- eqToRaw moved to module scope
- estimateTokens includes tool definitions
- splitMessage filters empty parts
- History save: SHA-256 diff check before DB write
- Cleanup interval stored and exportable

bot.ts (12):
- Pending maps: max-size enforcement (5000)
- pendingPublish fallthrough: console.warn
- _wizardLock cleared in periodic cleanup
- Marketplace: deactivate listing if seller agent deleted
- API key patterns extracted to shared constant
- Workflow templates: slug-based IDs instead of array indices
- API key encryption (AES-256-GCM) for storage/retrieval
- Per-user rate limiting (10 ops/min) on creation/edit/publish
- TON address validation via Address.parseFriendly()

Misc files (14):
- payments trackGeneration: refresh TTL on count update
- api-server: cleanup interval .unref()
- fragment-service: max-size caps on authStates + giftCache
- creator: removed dead schedule patterns
- notifier: _blockedUsers cap at 5000
- universal-agent-chat: XML tags for code injection prevention
- claude-code-bridge: 5MB stdout cap
- context: OWNER_ID parsed once at module level

+++ apps/builder-bot/src/db/schema-extensions.ts
@@ -169,8 +169,8 @@ export async function runMigrations(pool: Pool): Promise<void> {
         id            SERIAL PRIMARY KEY,
         user_id       BIGINT NOT NULL,
         type          TEXT NOT NULL,
-        amount_ton    DOUBLE PRECISION NOT NULL,
-        balance_after DOUBLE PRECISION NOT NULL DEFAULT 0,
+        amount_ton    NUMERIC(20,9) NOT NULL,
+        balance_after NUMERIC(20,9) NOT NULL DEFAULT 0,
         description   TEXT,
         tx_hash       TEXT,
         status        TEXT NOT NULL DEFAULT 'completed',
@@ -223,7 +223,11 @@ export async function runMigrations(pool: Pool): Promise<void> {
       )
     `);
 
-    // agent_approvals (AI agent action approvals)
+    // agent_approvals V1 DDL — original schema (action_type, action_details columns).
+    // V2 Drizzle definition (agentApprovalsTable at line ~948) uses different column names
+    // (action, description, details). The runAIProposalsMigrations() ALTER TABLE statements
+    // below ensure both column sets exist on the same table for backward compatibility.
+    // TODO: consolidate into a single schema once V1 callers are migrated.
     await client.query(`
       CREATE TABLE IF NOT EXISTS builder_bot.agent_approvals (
         id             SERIAL PRIMARY KEY,
@@ -945,6 +949,10 @@ export const agentAuditLogTable = builderSchema.table('agent_audit_log', {
 });
 
 // ─── Таблица 11: agent_approvals — запросы на одобрение действий ─────────
+// V2 Drizzle definition with (action, description, details) columns.
+// The V1 DDL in runMigrations() creates the table with (action_type, action_details).
+// ALTER TABLE statements in runAIProposalsMigrations() add both column sets so both work.
+// See also agentApprovalsTableV2 below and AgentApprovalsRepository which uses raw SQL.
 export const agentApprovalsTable = builderSchema.table('agent_approvals', {
   id:          serial('id').primaryKey(),
   agentId:     integer('agent_id').notNull(),
